requirements.txt
# FastAPI関連
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0

# LangChain関連
langchain>=0.1.0
langchain-openai>=0.0.2
langchain-community>=0.0.10

# Azure OpenAI関連
openai>=1.6.1,<2.0.0

# ベクトルストア関連
faiss-cpu==1.7.4
chromadb==0.4.18

# テキスト処理関連
tiktoken==0.5.2
sentence-transformers==2.2.2

# 負荷テスト関連
locust==2.17.0

# その他の依存関係
python-multipart==0.0.6
python-dotenv==1.0.0
httpx==0.25.2
aiofiles==23.2.1
numpy==1.24.3
pandas==2.1.4

# 開発・デバッグ用
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0

# モニタリング・分析用
wandb>=0.16.0
weave>=0.1.0
